chore: Update tests
