+++ src/test/java/com/google/cloud/tools/appengine/operations/DevServerJava8Test.java
@@ -17,7 +17,6 @@
 package com.google.cloud.tools.appengine.operations;
 
 import static com.google.common.base.StandardSystemProperty.JAVA_SPECIFICATION_VERSION;
-import static org.junit.Assert.assertThrows;
 import static org.junit.Assume.assumeTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -540,18 +539,5 @@ public class DevServerJava8Test {
   @Test
   public void testGetGaeRuntimeJava_isNotJava8() {
     Assert.assertEquals("java7", DevServer.getGaeRuntimeJava(false));
-  }
-
-  @Test
-  public void testInvalidJDKVersion_throwsIllegalArgumentException() {
-    RunConfiguration configuration =
-        RunConfiguration.builder(ImmutableList.of(java8Service)).projectJdkVersion("abc").build();
-    assertThrows(IllegalArgumentException.class, () -> devServer.run(configuration));
-
-    RunConfiguration configuration1 = configuration.toBuilder().projectJdkVersion("as37.5").build();
-    assertThrows(IllegalArgumentException.class, () -> devServer.run(configuration1));
-
-    RunConfiguration configuration2 = configuration.toBuilder().projectJdkVersion("1.b8c").build();
-    assertThrows(IllegalArgumentException.class, () -> devServer.run(configuration2));
   }
 }
