+++ src/test/java/com/google/cloud/tools/appengine/operations/DevServerJava9OrAboveTest.java
@@ -17,6 +17,7 @@
 package com.google.cloud.tools.appengine.operations;
 
 import static com.google.common.base.StandardSystemProperty.JAVA_SPECIFICATION_VERSION;
+import static org.junit.Assert.assertThrows;
 import static org.junit.Assume.assumeTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -891,5 +892,32 @@ public class DevServerJava9OrAboveTest {
   @Test
   public void testGetGaeRuntimeJava_isNotJava8() {
     Assert.assertEquals("java7", DevServer.getGaeRuntimeJava(false));
+  }
+
+  @Test
+  public void testValidJDKVersion() throws AppEngineException {
+    RunConfiguration configuration =
+        RunConfiguration.builder(ImmutableList.of(java8Service)).projectJdkVersion("1.8").build();
+    devServer.run(configuration);
+
+    RunConfiguration configuration1 =
+        RunConfiguration.builder(ImmutableList.of(java8Service)).projectJdkVersion("9").build();
+    devServer.run(configuration1);
+  }
+
+  @Test
+  public void testInvalidJDKVersion_throwsIllegalArgumentException() {
+    RunConfiguration configuration =
+        RunConfiguration.builder(ImmutableList.of(java8Service)).projectJdkVersion("12abc").build();
+    assertThrows(IllegalArgumentException.class, () -> devServer.run(configuration));
+
+    RunConfiguration configuration1 = configuration.toBuilder().projectJdkVersion("as37.5").build();
+    assertThrows(IllegalArgumentException.class, () -> devServer.run(configuration1));
+
+    RunConfiguration configuration2 = configuration.toBuilder().projectJdkVersion("1.a8s").build();
+    assertThrows(IllegalArgumentException.class, () -> devServer.run(configuration2));
+
+    RunConfiguration configuration3 = configuration.toBuilder().projectJdkVersion("12.42").build();
+    assertThrows(IllegalArgumentException.class, () -> devServer.run(configuration3));
   }
 }
